feat(inbox): retire RequestsApp + ReviewQueueKanban surfaces (Phase 2b)
Phase 2 (#802) collapsed tasks + requests + reviews into one unified
Inbox at /inbox. The legacy RequestsApp and ReviewQueueKanban surfaces
stayed wired in RootRoute.tsx alongside the new Inbox so deep links
kept working through the transition. This is the cleanup pass:

- Delete RequestsApp.tsx + RequestsApp.test.tsx.
- Delete ReviewQueueKanban.tsx + ReviewQueueKanban.test.tsx.
- Move the reusable RequestItem renderer to
  components/lifecycle/RequestItem.tsx (the unified inbox detail
  pane uses it; nothing else does).
- Wire /apps/requests and /reviews through a small InboxRedirect
  stub that pushes the user to /inbox so existing bookmarks keep
  working. Marked with data-testid="legacy-redirect-inbox" so the
  E2E suite can detect the redirect cleanly.
- DecisionInbox.tsx now imports RequestItem from the new location.

The two route IDs ("requests" / "reviews") are intentionally left in
routeRegistry so deep links remain resolvable; only the panels they
render were swapped. A future cleanup can drop the route IDs once
analytics confirms no more traffic.

bunx tsc --noEmit clean, web vitest 1516 pass.
